+++ src/main/java/peperact/common/block/peperact/BlockPeperact.java
@@ -1,9 +1,11 @@
 package peperact.common.block.peperact;
 
 import net.minecraft.block.ITileEntityProvider;
+import net.minecraft.block.SoundType;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
@@ -12,6 +14,7 @@ import net.minecraft.world.World;
 import net.minecraftforge.fml.common.registry.GameRegistry;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+import peperact.Peperact;
 import peperact.common.block.BlockBase;
 
 import javax.annotation.Nullable;
@@ -21,6 +24,9 @@ public class BlockPeperact extends BlockBase implements ITileEntityProvider {
 
     public BlockPeperact() {
         super(Material.IRON, MapColor.GREEN);
+        this.setHardness(5.0F);
+        this.setSoundType(SoundType.METAL);
+        this.setCreativeTab(Peperact.CREATIVE_TAB);
         // Don't know if I should be registering it here... XD
         GameRegistry.registerTileEntity(TilePeperact.class, this.getResourceLocation());
     }